Integrate Blender deck simulation helpers
