+++ simulations/sphere_space_station_simulations/blender_helpers/README.md
@@ -1,0 +1,3 @@
+# Blender Helpers
+
+Utility functions shared by the Blender adapter scripts.
